% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denominator-of-clustfun.R
\name{clustfun_denominator}
\alias{clustfun_denominator}
\title{Estimate the denominator of clustering function}
\usage{
clustfun_denominator(x, r, correction = "best", ...)
}
\arguments{
\item{x}{Point pattern}

\item{r}{Vector of distances to estimate the function}

\item{correction}{Border correction. Either "border", "none" or "best"(="border).}

\item{...}{Ignored.}
}
\value{
\code{\link{fv}}-object.
}
\description{
Estimate the denominator of clustering function
}
\details{
Support in 3D therefore only for cuboidal windows. (inc. rotated)
}
