% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalise_pp.R
\name{internalise_pp}
\alias{internalise_pp}
\title{Convert input to SGCS internal point pattern data}
\usage{
internalise_pp(x)
}
\arguments{
\item{x}{some form of point pattern data.}
}
\description{
Convert input to SGCS internal point pattern data
}
\details{
Understands 'ppp', 'pp3', 'matrix' and list(coordinate_matrix, bbox).

Data.frame marks are not well understood, so convert the mark data to
suit the analysis before use. 
 
Mainly for internal use.
}
