% Generated by roxygen2 (4.0.2): do not edit by hand
\name{morphoEuler}
\alias{morphoEuler}
\title{Morphologicals: Euler number of dilated pattern}
\usage{
morphoEuler(x, r, ...)
}
\arguments{
\item{x}{Point pattern}

\item{r}{Vector of distances to estimate the function}

\item{...}{Ignored.}
}
\value{
Only for 2D.

Reduced sample border correction.
}
\description{
Dilate pattern with b(o,r) and compute the Euler number.
}

