% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morpho-Length.R
\name{morphoLength}
\alias{morphoLength}
\title{Morphologicals: Relative boundary length of diluted pattern}
\usage{
morphoLength(x, r, ...)
}
\arguments{
\item{x}{Point pattern}

\item{r}{Vector of distances to estimate the function}

\item{...}{Ignored.}
}
\value{
Only for 2D.

The default plotted curve, "u(r)" is U(r)/(2 * r * lambda * pi)
with U(r) the boundary length fraction.

Reduced sample border correction.
}
\description{
Dilute pattern with b(o,r) and compute boundary length.
}
