% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pairwise_distances}
\alias{pairwise_distances}
\title{pairwise distances}
\usage{
pairwise_distances(x, toroidal = FALSE)
}
\arguments{
\item{x}{Point pattern data in some format.}

\item{toroidal}{For toroidal distances.}
}
\value{
Value is the lower triangle of the distasnce matrix, returned as a vector.
}
\description{
pairwise distances
}

