% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Rfun}
\alias{Rfun}
\title{Clustering function versio 2}
\usage{
Rfun(x, r, correction = "border", scaled = FALSE, ...)
}
\arguments{
\item{x}{Point pattern}

\item{r}{Vector of distances}

\item{correction}{Border correction. "none" or "border" (reduced window) supported.}

\item{scaled}{Scale with theoretical value?}

\item{...}{ignored.}
}
\value{
\code{\link{fv}}-object, see \link{spatstat} for more.
}
\description{
Ratio of T and K
}
\details{
Reduced border correction available.
}
\examples{
\dontrun{
en <- envelope(rcell(nx=15), fun=Rfun)
plot(en)
}
}

