% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/T-function.R
\name{Tfun}
\alias{Tfun}
\title{Triplet intensity function}
\usage{
Tfun(x, r, ...)
}
\arguments{
\item{x}{Point pattern}

\item{r}{Vector of distances to estimate the function}

\item{...}{ignored.}
}
\value{
\code{\link{fv}}-object.
}
\description{
Summarise the number of r-close triangles in a stationary and isotropic point pattern (2d,3d).
}
\details{
Border correction is done via minus sampling.

See 

\strong{Schladitz, Baddeley: A Third order point process characteristic, SJS, vol 27, 657-671, 2000.}

for details.
}
