% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bounding-box-xy.R
\name{bounding_box_xy}
\alias{bounding_box_xy}
\title{Compute the bounding window from coordinates}
\usage{
bounding_box_xy(x)
}
\arguments{
\item{x}{Columnwise matrix of coordinates, n-dimensional.}
}
\description{
Compute the bounding window from coordinates
}
\details{
Rectangular or cuboidal bounding box around coordinates given as
matrix 'x'. Enlarged with the Ripley and Rasson 1977 method.
}
