% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{default_r}
\alias{default_r}
\title{Default range vector}
\usage{
default_r(x, r)
}
\arguments{
\item{x}{point pattern, internal format}

\item{r}{most likely missing. if not , returned.
Compute the range vector}
}
\description{
Default range vector
}
