% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge-distance.R
\name{edge_distance}
\alias{edge_distance}
\title{Distances to observation window edge}
\usage{
edge_distance(x)
}
\arguments{
\item{x}{Point pattern

Uses bdist.points for ppp objects, otherwise relies on a cuboidal/rectangular window.}
}
\description{
Compute the distances from points to observation window edge.
}
