% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/K-function.R
\name{Kfun}
\alias{Kfun}
\title{Ripley's K-function}
\usage{
Kfun(x, r, ...)
}
\arguments{
\item{x}{Point pattern}

\item{r}{Vector of distances to estimate the function}

\item{...}{ignored.}
}
\value{
\code{\link{fv}}-object.
}
\description{
See \code{\link[spatstat.core]{Kest}} for more rigorous implementation in 2D.
}
\details{
Border correction is translation correction. 
Support in 3D therefore only for cuboidal windows.
}
