% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R-function.R
\name{Rfun}
\alias{Rfun}
\title{Clustering function versio 2}
\usage{
Rfun(x, r, correction = "border", scaled = FALSE, ...)
}
\arguments{
\item{x}{Point pattern}

\item{r}{Vector of distances}

\item{correction}{Border correction. "none" or "border" (reduced window) supported.}

\item{scaled}{Scale with theoretical value?}

\item{...}{ignored.}
}
\value{
\code{\link{fv}}-object, see \link{spatstat} for more.
}
\description{
Ratio of triplets to pairs^2, motivated by T/K^2.
}
\details{
Reduced border correction available.
}
\examples{
\dontrun{
en <- envelope(rcell(nx=15), fun=Rfun)
plot(en)
}
}
