% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translation-weights.R
\name{translation_weights}
\alias{translation_weights}
\title{Translation weights
Compute the translation edge correction weights}
\usage{
translation_weights(x)
}
\arguments{
\item{x}{Some form of point pattern data.}
}
\description{
Translation weights
Compute the translation edge correction weights
}
